+++ paper/paper.bib
@@ -1,32 +1,59 @@
 @book{MacKey:2003,
-  	Author = {David MacKay},
-  	Publisher = {Cambridge University Press},
-  	Title = {{Information Theory, Inference, and Learning Algorithms}},
-  	Year = 2003
+  title = "Information theory, inference and learning algorithms",
+  author = "MacKay, David J C",
+  publisher = "Cambridge University Press",
+  month = sep,
+  year = 2003,
+  address = "Cambridge, England",
+  language = "en",
+  doi = {10.1109/tit.2004.834752}
 }
 
-@book{Shannon:1948,
-  	Author = {Shannon, Claude E.},
-  	Publisher = {Bell System Technical Journal. 27},
-  	Title = {{A Mathematical Theory of Communication}},
-    month = jul,
-  	Year = 1948
+@article{Shannon:1948,
+  title = "A mathematical theory of communication",
+  author = "Shannon, C E",
+  journal = "Bell Syst. tech. j.",
+  publisher = "Institute of Electrical and Electronics Engineers (IEEE)",
+  volume = 27,
+  number = 3,
+  pages = "379--423",
+  month = jul,
+  year = 1948,
+  doi = {10.1002/j.1538-7305.1948.tb01338.x}
 }
 
 @article{Huffman:1952,
-    author = {D. A. Huffman},
-    title = "{A method for the construction of minimum-redundancy codes}",
-    journal = {Proc. IRE, vol. 40, pp. 1098-1101},
-    year = 1952,
-    volume = 40
+  title = "A method for the construction of minimum-redundancy codes",
+  author = "Huffman, David",
+  journal = "Proc. IRE",
+  publisher = "Institute of Electrical and Electronics Engineers (IEEE)",
+  volume = 40,
+  number = 9,
+  pages = "1098--1101",
+  month = sep,
+  year = 1952,
+  doi = {10.1109/JRPROC.1952.273898}
 }
 
 @article{LempelZiv:1977,
-    author = {J. Ziv and A. Lempel},
-    title = "{A Universal Algorithm for Data Compression}",
-    journal = {IEEE Transactions on Information Theory 23(3):337–343},
+  title = "A universal algorithm for sequential data compression",
+  author = "Ziv, J and Lempel, A",
+  journal = "IEEE Trans. Inf. Theory",
+  publisher = "Institute of Electrical and Electronics Engineers (IEEE)",
+  volume = 23,
+  number = 3,
+  pages = "337--343",
+  month =  may,
+  year = 1977,
+  doi = {10.1109/TIT.1977.1055714}
+}
+
+@techreport{BurrowsWheeler:1994,
+    author = {M. Burrows and D. Wheeler},
+    title = "{A Block-Sorting Lossless Data Compression Algorithm}",
+    journal = {Research Report 124, Digital Equipment Corporation, Palo Alto, CA, USA},
     month = may,
-    year = 1977
+    year = 1994
 }
 
 @misc{Delta_encoding:2019,
@@ -35,14 +62,6 @@
   year = {2019},
   publisher = {Wikipedia},
   url = {https://en.wikipedia.org/wiki/Delta_encoding}
-}
-
-@article{BurrowsWheeler:1994,
-    author = {M. Burrows and D. Wheeler},
-    title = "{A Block-Sorting Lossless Data Compression Algorithm}",
-    journal = {Research Report 124, Digital Equipment Corporation, Palo Alto, CA, USA},
-    month = may,
-    year = 1994
 }
 
 @misc{StatisticalDistribution:2017,
