### 2021-11-05 #3890 @siara-cc
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):

I have added DOIs where ever available
