### 2021-11-16 #3919 @whedon
[REVIEW]: Unishox: A hybrid encoder for Short Unicode Strings
**Submitting author:** @siara-cc (<a href="http://orcid.org/0000-0001-6642-447X">Arundale Ramanathan</a>)
**Repository:** <a href="https://github.com/siara-cc/Unishox2" target ="_blank">https://github.com/siara-cc/Unishox2</a>
**Version:** 1.0.2
