### 2021-11-17 #3919 @gradvohl
# Reviewer report for JOSS paper #3919
**Paper title**: Unishox: A hybrid encoder for Short Unicode Strings

## Review criteria
* A list of the authors of the software: :heavy_check_mark:
* Author affiliations: :heavy_check_mark: (The author identifies himself as an "*Independent Researcher*").
* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience: :heavy_check_mark: (Please see the [**Suggestions in the *Summary***](#suggestions-in-the-summary) section in the [**Reviewer Comments**](#reviewer-comments) section)
* A clear statement of need that illustrates the purpose of the software: :heavy_check_mark:
* Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it: :heavy_check_mark:  (Please see the [**Some additional information in the paper**](#some-additional-information-in-the-paper) section in the [**Reviewer Comments**](#reviewer-comments) section).
* A list of key references including a link to the software archive:  :wave: (sort of -- Please see the [**References in the paper**](#references-in-the-paper) section in the [**Reviewer Comments**](#reviewer-comments) section)

## Review items
* Software license: :thumbsup: (good -- Apache-2.0 License)
* Substantial scholarly effort: :thumbsup: (good -- Please see the [**Scholarly effort**](#scholarly-effort)  section in the [**Reviewer Comments**](#reviewer-comments) section)
* Documentation:
    * A statement of need: :thumbsup: (good)
    * Installation instructions: :thumbsup: (good)
    * Example usage: :thumbsup: (good)
    * API documentation: :thumbsdown: (**bad** -- Please see the [**API Docs**](#api-docs)  section in the [**Reviewer Comments**](#reviewer-comments) section)
    * Community guidelines:  :thumbsdown: (**bad** -- Please see the [**Community guidelines**](#community-guidelines)  section in the [**Reviewer Comments**](#reviewer-comments) section)
    * Functionality: :thumbsup: (good)
    * Tests: :thumbsup: (good)

* Software paper
    * Authors: :heavy_check_mark:
    * A statement of need: :heavy_check_mark:
    * Quality of writing: :wave: (sort of -- A little improvement in writing would be good. Perhaps an English native speaker may give interesting suggestions).
    * References: :heavy_check_mark: (Please see the [**References in the paper**](#references-in-the-paper)  section in the [**Reviewer Comments**](#reviewer-comments) section)

## Test environment
The reviewer conducted the tests for this software using an AWS instance (t2.micro) with Amazon Linux 2.

## Reviewer comments
The following comments refer to items I considered  that could be improved, although they currently met the requirements for JOSS submission.

### Suggestions in the *Summary*
In the *Summary* section of the paper, the author claims that the "hybrid encoding method for compressing Short Unicode Strings of arbitrary lengths including Latin/English text and printable special characters (...) has not been sufficiently addressed by lossless encoding methods so far." It would be nice if he included an updated reference to support this statement. After all, most of the citations in the paper are more than five years old.

### Some additional information in the paper
The ``README.md`` file in the repository contains information about other projects that use ``Unishox``. However, those mentions are not in the ``paper.md`` file. It would be good if the author incorporated those mentions also in the ``paper.md`` file. Perhaps, in a specific section.

### References in the Paper
Please, add the link for the work of Burrows & Wheeler (1994) https://www.hpl.hp.com/techreports/Compaq-DEC/SRC-RR-124.pdf.

### Scholarly effort
I consider that there is a substantial scholarly effort producing this code. However, it would be nice if the author publishes a paper discussing the features and advantages of the ``Unishox2`` compared with other algorithms. Also, a publication in a scientific journal or a conference would be interesting. However, if the paper is on a preprint server (such as the Computer Science section of [arXiv](https://arxiv.org)) it is a significant first step.

### API Docs
I did not find any documentation about the code regarding the API documentation, neither in the ``docs`` directory (it contains only a ``.tex`` file), nor in the source code. Therefore, I strongly recommend the documentation of the source code using a standard.
My suggestion is to use the [Doxygen](https://www.doxygen.nl/index.html) software. It can generate the docs (in HTML or PDF) from the documented source files.

### Community guidelines
Looking for information about how to contribute to Unishox2 software, I did not find any guidelines. There is only the **Issues** section in the repository, where there is the author's email to report issues or seek support. My suggestion is to create a **Contributing** section (or a ``CONTRIBUTING.md`` file) with more specific instructions about how to contribute.

### Other Suggestions
When I compiled with an older GCC version (``4.8.5``), the compiler issued the following error: ``‘for’ loop initial declarations are only allowed in C99 mode``. To avoid such problems, please inform the parameter ``-std=c99`` to the ``gcc`` compiler in the ``Makefile``.

Please, consider including the ``CITATION.cff`` file with information about how to cite the ``Unishox`` software. In addition, information about software citation in GitHub is available at [GitHub Docs](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files).

## Conclusion
Analyzing all the aspects mentioned earlier, I rank the paper as **major revisions** needed. Although most of those revisions are very easy to implement, I consider that a lack of documentation prevents acceptance of the paper at this time.
