+++ paper/paper.bib
@@ -135,3 +135,31 @@
     year = {2015},
     url = {https://ed-von-schleck.github.io/shoco}
 }
+
+@article{Gardner:2013,
+    author = {{Gardner-Stephen et. al}},
+    year = {2013},
+    month = {12},
+    pages = {497},
+    title = {Improving Compression of Short Messages},
+    volume = {6},
+    journal = {International Journal of Communications, Network and System Sciences},
+    doi = {10.4236/ijcns.2013.612053}
+}
+
+@article{Xu2017SSELC:2017,
+  title={SSE Lossless Compression Method for the Text of the Insignificance of the Lines Order},
+  author={{Juncai Xu et. al}},
+  journal={ArXiv},
+  year={2017},
+  volume={abs/1709.04035}
+}
+
+@INPROCEEDINGS{AIMCS:2020,
+  author={Abedi, Masoud and Pourkiani, Mohammadreza},
+  booktitle={2020 IEEE 18th World Symposium on Applied Machine Intelligence and Informatics (SAMI)}, 
+  title={AIMCS: An Artificial Intelligence based Method for Compression of Short Strings}, 
+  year={2020},
+  pages={311-318},
+  doi={10.1109/SAMI48414.2020.9108719}
+}
