+++ paper/paper.md
@@ -21,11 +21,17 @@ bibliography: paper.bib
 
 Unishox is a hybrid encoding technique with which short unicode strings could be compressed using context aware pre-mapped codes and delta coding resulting in surprisingly good ratios.
 
-This article discusses a hybrid encoding method for compressing Short Unicode Strings of arbitrary lengths including Latin/English text and printable special characters. This has not been sufficiently addressed by lossless encoding methods so far.
+This article discusses a hybrid encoding method for compressing Short Unicode Strings of arbitrary lengths including Latin/English text and printable special characters. This has not been sufficiently addressed by lossless short text encoding methods so far as they have one or more of the following drawbacks:
+
+- They apply only to either English text or Unicode strings and not both
+- They have specific criteria based on which best compression can be achieved 
+- They are not suitable for low RAM devices
+
+To the extent we know, the existing methods available for short string compression are Smaz [@Sanfilippo:2012], Shoco [@Schramm:2015], SCSU, [@Unicode:2005], BOCU [@Scherer:2002], SSE [@Xu2017SSELC:2017] and AIMCS [@AIMCS:2020].
 
 # Statement of need
 
-Space occupied by short strings become significant in memory constrained environments such as Arduino Uno and ESP8266. Text exchange in Chat applications is another area where cost savings could be seen using such compression. It is also possible to achieve savings in bandwidth and storage cost by storing and retrieving independent strings in Cloud databases.
+Space occupied by short strings become significant in memory constrained environments such as Arduino Uno and ESP8266. Text exchange in Chat applications and social media posts is another area where cost savings could be seen using such compression. It is also possible to achieve savings in bandwidth and storage cost by storing and retrieving independent strings in Cloud databases.
 
 # Existing Techniques
 
@@ -45,9 +51,21 @@ The LZ77 family of encoders use the dictionary encoding technique for compressin
 
 In contrast to these encoding methods, there are various other approaches to lossless coding including Run Length Encoding (RLE) and Burrows-Wheeler coding [@BurrowsWheeler:1994].
 
-While technologies such as GZip, Deflate, Zip, LZMA are available for general purpose compression, they do not provide optimal compression for short strings. Eventhough these methods compress far more than what is proposed in this article, they often expand the original source for short strings because the symbol-code mapping also needs to be attached to aid decompression.
+While programs such as GZip, Deflate, Zip, LZMA and BZ2 that use such technologies are available for general purpose compression, they do not provide optimal compression for short strings. Eventhough these methods compress far more than what is proposed in this article, they often expand the original source for short strings because the symbol-code mapping also needs to be attached to aid decompression.
 
-For compressing Unicode sequences, three other technologies are available: SCSU [@Unicode:2005] and BOCU [@Scherer:2002] and FAST [@Studený:2008]. A survey of these compression algorithms was published by Doug Ewell in 2004 [@Ewell:2004].
+## Short string compression techniques
+
+Techniques available For compressing short English / Latin text are Smaz and shoco, but are not developed with Unicode in mind.
+
+Smaz is a simple compression library suitable for compressing very short strings [@Sanfilippo:2012]. It was developed by Salvatore Sanfilippo and is released under the BSD license.
+
+Shoco is a C library to compress short strings [@Schramm:2015]. It was developed by Christian Schramm and is released under the MIT license.
+
[…]
