+++ html/search/functions_4.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['presetforunicode_0',['presetForUnicode',['../test__unishox2_8c.html#a87b6ce76f87c12a590ac61cc52311cbe',1,'test_unishox2.c']]],
+  ['print_5fcompressed_1',['print_compressed',['../test__unishox2_8c.html#a0157c73cbed609f4961b2d4f57607231',1,'test_unishox2.c']]],
+  ['print_5fstring_5fas_5fhex_2',['print_string_as_hex',['../test__unishox2_8c.html#aa10ed420988aa89f748155b9789fdb30',1,'test_unishox2.c']]]
+];
