### 2021-11-20 #3919 @siara-cc
@gradvohl I have addressed all the points raised during review and made commits to the repo.  Please review to see if the changes meet expected standards when it is convenient.  I have also added the pre-review latex pdf as suggested to ResearchGate [at this url](https://www.researchgate.net/publication/356411501_Unishox_-A_hybrid_encoder_for_Short_Unicode_Strings) since arXiv is off-limits for an independent researcher like me.

Thanks again for spending time on this and the significant observations.
