### 2021-11-20 #3919 @gradvohl
Hello @siara-cc

Sorry, but not all points were appropriately addressed. In my opinion, there are some missed points as follows:
1. The link for Burrows & Wheeler (1994)'s work in the reference section is still missing.
2. There are no comments on the source code using a standard (e.g., Doxygen or other software).
3. The CONTRIBUTING.md file misses the "project team" email (actually your email) in the "**Enforcement**" section.
4. To maintain compatibility with the new ``Makefile``, adding the ``-std=c99`` flag in the **Compiling** section in the ``README.md`` file would be significant.

Regarding the pre-review file, I see no problems publishing it on ResearchGate.net (although I prefer [Figshare](https://figshare.com) for this task). However, I am not sure if you did it right. The file is published at ResearchGate, indeed. Nevertheless, the DOI points to your repository (with the source files) and not to the paper itself. Take a look at the instructions to generate a DOI for your paper [here](https://explore.researchgate.net/display/support/ResearchGate+DOIs).

I believe that is it for now.

Best regards.
