### 2021-11-20 #3919 @gkthiruvathukal
@gradvohl  Thanks for your comments.  @siara-cc, can you please address these as soon as possible?

@siara-cc Can you also clarify your points about ArXiv? It should absolutely not be "off-limits". I checked the submission requirements again, and don't think you are in any way prevented from submitting. If you mean that you are not allowed to post, it is just a matter of having an endorser. @gradvohl's suggestion of figshare is also a great option. I personally find it a lot easier to use than ArXiv!

@linuxscout Can you let us know when you have a chance to review the submission and fill in the checklist?
