+++ paper/paper.bib
@@ -52,6 +52,7 @@
     author = {M. Burrows and D. Wheeler},
     title = "{A Block-Sorting Lossless Data Compression Algorithm}",
     journal = {Research Report 124, Digital Equipment Corporation, Palo Alto, CA, USA},
+    url = {https://www.hpl.hp.com/techreports/Compaq-DEC/SRC-RR-124.pdf},
     month = may,
     year = 1994
 }
@@ -152,6 +153,7 @@
   author={{Juncai Xu et. al}},
   journal={ArXiv},
   year={2017},
+  url = {https://arxiv.org/pdf/1709.04035},
   volume={abs/1709.04035}
 }
 
