+++ html/search/all_3.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['main_0',['main',['../test__unishox2_8c.html#a0ddf1224851353fc92bfbff6f499fa97',1,'test_unishox2.c']]]
-];
