+++ html/search/functions_2.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['gettimeval_0',['getTimeVal',['../test__unishox2_8c.html#af9ea2515c2ce8ef756cb24b8f82f5213',1,'test_unishox2.c']]]
-];
