+++ html/search/functions_7.js
@@ -1,11 +1,0 @@
-var searchData=
-[
-  ['unishox2_5fcompress_0',['unishox2_compress',['../unishox2_8h.html#a379f7637c1038582e5e3fcab27b4fae0',1,'unishox2_compress(const char *in, int len, UNISHOX_API_OUT_AND_LEN(char *out, int olen), const unsigned char usx_hcodes[], const unsigned char usx_hcode_lens[], const char *usx_freq_seq[], const char *usx_templates[]):&#160;unishox2.c'],['../unishox2_8c.html#af7f58e19f06f8e7d0d1bb4b750f1f7c4',1,'unishox2_compress(const char *in, int len, UNISHOX_API_OUT_AND_LEN(char *out, int olen), const byte usx_hcodes[], const byte usx_hcode_lens[], const char *usx_freq_seq[], const char *usx_templates[]):&#160;unishox2.c']]],
-  ['unishox2_5fcompress_5flines_1',['unishox2_compress_lines',['../unishox2_8h.html#a5fc0490b257381a887c27e6c986e76bc',1,'unishox2_compress_lines(const char *in, int len, UNISHOX_API_OUT_AND_LEN(char *out, int olen), const unsigned char usx_hcodes[], const unsigned char usx_hcode_lens[], const char *usx_freq_seq[], const char *usx_templates[], struct us_lnk_lst *prev_lines):&#160;unishox2.c'],['../unishox2_8c.html#a759f6131a89c84a206b7ce9d8bc73a78',1,'unishox2_compress_lines(const char *in, int len, UNISHOX_API_OUT_AND_LEN(char *out, int olen), const byte usx_hcodes[], const byte usx_hcode_lens[], const char *usx_freq_seq[], const char *usx_templates[], struct us_lnk_lst *prev_lines):&#160;unishox2.c']]],
-  ['unishox2_5fcompress_5fpreset_5flines_2',['unishox2_compress_preset_lines',['../test__unishox2_8c.html#a0f5a0f7e959c0ce34e4ccffb63bd0bb7',1,'test_unishox2.c']]],
-  ['unishox2_5fcompress_5fsimple_3',['unishox2_compress_simple',['../unishox2_8h.html#a37df146e9ddd9ff1b07c7c160683d96c',1,'unishox2_compress_simple(const char *in, int len, char *out):&#160;unishox2.c'],['../unishox2_8c.html#a37df146e9ddd9ff1b07c7c160683d96c',1,'unishox2_compress_simple(const char *in, int len, char *out):&#160;unishox2.c']]],
-  ['unishox2_5fdecompress_4',['unishox2_decompress',['../unishox2_8h.html#a1fd9aa9f14850e527e6b46089cf83d1a',1,'unishox2_decompress(const char *in, int len, UNISHOX_API_OUT_AND_LEN(char *out, int olen), const unsigned char usx_hcodes[], const unsigned char usx_hcode_lens[], const char *usx_freq_seq[], const char *usx_templates[]):&#160;unishox2.c'],['../unishox2_8c.html#a6e7c35fe5cf2bfab017580f19a1a2d01',1,'unishox2_decompress(const char *in, int len, UNISHOX_API_OUT_AND_LEN(char *out, int olen), const byte usx_hcodes[], const byte usx_hcode_lens[], const char *usx_freq_seq[], const char *usx_templates[]):&#160;unishox2.c']]],
[…]
