+++ html/search/all_5.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['run_5funit_5ftests_0',['run_unit_tests',['../test__unishox2_8c.html#a16080e386b8908144ddddf1ff49a4a31',1,'test_unishox2.c']]]
-];
