+++ html/search/functions_6.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['test_5fushx_5fcd_0',['test_ushx_cd',['../test__unishox2_8c.html#ab1d0d91638554951b3f7229cd2001916',1,'test_unishox2.c']]],
-  ['timedifference_1',['timedifference',['../test__unishox2_8c.html#ae759f85f0948372a6591ebb190a595c4',1,'test_unishox2.c']]]
-];
