+++ html/search/all_6.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['test_5funishox2_2ec_0',['test_unishox2.c',['../test__unishox2_8c.html',1,'']]],
-  ['test_5fushx_5fcd_1',['test_ushx_cd',['../test__unishox2_8c.html#ab1d0d91638554951b3f7229cd2001916',1,'test_unishox2.c']]],
-  ['timedifference_2',['timedifference',['../test__unishox2_8c.html#ae759f85f0948372a6591ebb190a595c4',1,'test_unishox2.c']]]
-];
