+++ html/search/files_0.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['test_5funishox2_2ec_0',['test_unishox2.c',['../test__unishox2_8c.html',1,'']]]
-];
