+++ html/unishox2_8c.js
@@ -1,9 +1,0 @@
-var unishox2_8c =
-[
-    [ "unishox2_compress", "unishox2_8c.html#af7f58e19f06f8e7d0d1bb4b750f1f7c4", null ],
-    [ "unishox2_compress_lines", "unishox2_8c.html#a759f6131a89c84a206b7ce9d8bc73a78", null ],
-    [ "unishox2_compress_simple", "unishox2_8c.html#a37df146e9ddd9ff1b07c7c160683d96c", null ],
-    [ "unishox2_decompress", "unishox2_8c.html#a6e7c35fe5cf2bfab017580f19a1a2d01", null ],
-    [ "unishox2_decompress_lines", "unishox2_8c.html#aec21ab3385bd5920e2ce6c63d1ebf418", null ],
-    [ "unishox2_decompress_simple", "unishox2_8c.html#a3e5c00a0c5db7979d94b0b294908a04e", null ]
-];

[…]
