### 2021-11-20 #3919 @siara-cc
Hi @gradvohl ,

Thank you so much for reviewing it again. I have addressed these issues and please see my replies:

> 1. The link for Burrows & Wheeler (1994)'s work in the reference section is still missing.

I realise now that the URL is missing and I have added it now.  If my understanding is incorrect, please let me know.

> 2. There are no comments on the source code using a standard (e.g., Doxygen or other software).

I have added Doxygen comments to `unishox2.c`, `unishox2.h` and `test_unishox2.c`, which are the source code files related to the present version of Unishox (byte code version 2).  However, I have generated Doxygen pages only for `unishox2.h`, which describes the API functions.

The generated Doxygen API documentation can be found at: https://siara-cc.github.io/Unishox2/unishox2_8h.html

I have also addressed points (3) and (4).

Further, I have added the paper to FigShare [here](https://figshare.com/articles/preprint/Unishox_A_hybrid_encoder_for_Short_Unicode_Strings/17056334) with a DOI as suggested.  I also removed the one having incorrect DOI from ResearchGate.

As far as I have checked, I have addressed all the points.  If I have missed something or in case of other suggestions, please let me know.  I really appreciate the detailed review and advice. It will go a long way towards improving my future submissions.
