### 2021-11-20 #3919 @gkthiruvathukal
@siara-cc Figshare is fine, for now. If you continue to publish and want to use ArXiv, you can reach out for an endorsement. I've had people ask me in the past. Figshare will do for now!
