### 2021-11-21 #3919 @gradvohl
Hello @siara-cc,

I apologize for being so critical, but I am still not satisfied with some information in your repository.

First, I understand that you prefer to document the functions in the `.h` file, and I see no problems in doing this. However, in the "**File List**" section on the generated documentation, there is no reference to the ``unishox2.c`` file. Therefore, since you are expecting to publish in an **Open Source** Journal, I think that the documentation should point to the file that actually implements the code. So, please, include a reference to the ``unishox2.c`` file in the "**File List**" on the Doxygen generated documentation.

Second, regarding the [file uploaded to the Figshare](https://figshare.com/articles/preprint/Unishox_A_hybrid_encoder_for_Short_Unicode_Strings/17056334), when we click on the link you informed, it takes us to the Figshare site. However, the site reports that "``Linked content is NOT stored on Figshare, and we can't guarantee its availability, quality, security or accept any liability.``" and points us back to the paper your GitHub repository. Unfortunately, that is not what a reader expects.

Having a DOI implies that the file (your paper in this case) is available right when we click on the DOI. Therefore, please, make your file ready for download (or read) in Figshare with a single click. In addition, as a suggestion, inform the reader about the paper in your main ``README.md`` file, so [&epsilon;|s]he can understand the theory/logic implemented in the code.

Best.
