+++ unishox2.c
@@ -33,9 +33,13 @@
 
 #include "unishox2.h"
 
+/// byte is unsigned char
 typedef unsigned char byte;
 
+/// possible horizontal sets and states
 enum {USX_ALPHA = 0, USX_SYM, USX_NUM, USX_DICT, USX_DELTA, USX_NUM_TEMP};
+
+/// This 2D array has the characters for the sets USX_ALPHA, USX_SYM and USX_NUM. Where a character cannot fit into a byte, 0 is used and handled in code.
 byte usx_sets[][28] = {{  0, ' ', 'e', 't', 'a', 'o', 'i', 'n',
                         's', 'r', 'l', 'c', 'd', 'h', 'u', 'p', 'm', 'b',
                         'g', 'w', 'f', 'y', 'v', 'k', 'q', 'j', 'x', 'z'},
@@ -46,49 +50,78 @@ byte usx_sets[][28] = {{  0, ' ', 'e', 't', 'a', 'o', 'i', 'n',
                         '/', '3', '4', '6', '7', '8', '(', ')', ' ',
                         '=', '+', '$', '%', '#', 0, 0, 0, 0, 0}};
 
-// Stores position of letter in usx_sets.
-// First 3 bits - position in usx_hcodes
-// Next  5 bits - position in usx_vcodes
+/// Stores position of letter in usx_sets.
+/// First 3 bits - position in usx_hcodes
+/// Next  5 bits - position in usx_vcodes
 byte usx_code_94[94];
 
+/// Vertical codes starting from the MSB
 byte usx_vcodes[]   = { 0x00, 0x40, 0x60, 0x80, 0x90, 0xA0, 0xB0,
                         0xC0, 0xD0, 0xD8, 0xE0, 0xE4, 0xE8, 0xEC,
                         0xEE, 0xF0, 0xF2, 0xF4, 0xF6, 0xF7, 0xF8,
                         0xF9, 0xFA, 0xFB, 0xFC, 0xFD, 0xFE, 0xFF };
+
+/// Length of each veritical code
 byte usx_vcode_lens[] = {  2,    3,    3,    4,    4,    4,    4,
                            4,    5,    5,    6,    6,    6,    7,
                            7,    7,    7,    7,    8,    8,    8,
                            8,    8,    8,    8,    8,    8,    8 };
 
+/// Vertical Codes and Set number for frequent sequences in sets USX_SYM and USX_NUM. First 3 bits indicate set (USX_SYM/USX_NUM) and rest are vcode positions
 byte usx_freq_codes[] = {(1 << 5) + 25, (1 << 5) + 26, (1 << 5) + 27, (2 << 5) + 23, (2 << 5) + 24, (2 << 5) + 25};
 
+/// Not used
 const int UTF8_MASK[] = {0xE0, 0xF0, 0xF8};
+/// Not used
 const int UTF8_PREFIX[] = {0xC0, 0xE0, 0xF0};
 
+/// Minimum length to consider as repeating sequence
 #define NICE_LEN 5
 
+/// Set (USX_NUM - 2) and vertical code (26) for encoding repeating letters
 #define RPT_CODE ((2 << 5) + 26)
+/// Set (USX_NUM - 2) and vertical code (27) for encoding terminator
 #define TERM_CODE ((2 << 5) + 27)
+/// Set (USX_SYM - 1) and vertical code (7) for encoding Line feed \\n
 #define LF_CODE ((1 << 5) + 7)
+/// Set (USX_NUM - 1) and vertical code (8) for encoding \\r\\n
 #define CRLF_CODE ((1 << 5) + 8)
+/// Set (USX_NUM - 1) and vertical code (22) for encoding \\r
[…]
