### 2021-11-21 #3919 @siara-cc
Hi @gradvohl , Sure, I have added `unishox2.c` to Doxygen doc and uploaded the actual PDF to figshare with a link in README.md.  I have documented each function and variable in `unishox2.c` to make it easy to understand for viewers and future contributors.  Thank you for pointing this out.
