### 2021-11-22 #3919 @siara-cc
Hi @gradvohl it still shows 2 items pending for the first review. Anything I need to do to close them ?
@gkthiruvathukal
