### 2021-11-22 #3919 @gradvohl
Hello @siara-cc,
Your software does not cover the two items (Automated Tests and Performance claims). That is the reason why I left both items checked off.
