### 2021-11-22 #3919 @siara-cc
Hi @gradvohl I have unit tests that provide good coverage and presented performance numbers in the performance section comparing to other competitive technologies. Also another person has made comprehensive tests and performance comparison with reddit posts as mentioned in the credits section. Kindly see if this meets thess requirements.
