### 2021-11-22 #3919 @gkthiruvathukal
@gradvohl is right. In some cases, the items are not required. Being a software engineer myself, I like to see automatic tests  or unit tests but there are some cases where requiring it doesn't make sense.

I'm still hoping to hear some input from @linuxscout. Based on the detailed review of @gradvohl and subsequent efforts by @siara-cc to resolve the issues, I feel like we are moving toward an acceptable submission.

Given that the US holiday of Thanksgiving is just around the corner, I'm going to suggest that we wait for @linuxscout's review. We're in pretty good time on this submission so I'd like to allow for plenty of time for this review to be submitted before proceeding with the final decision, but things are looking encouraging!
