### 2021-11-22 #3919 @siara-cc
Thanks @gkthiruvathukal but I have both unit tests and performance tests in place. Please see my reply here: https://github.com/openjournals/joss-reviews/issues/3919#issuecomment-976005023
