### 2021-11-23 #3919 @gradvohl
Hello @gkthiruvathukal and @siara-cc,
As far as I know, the performance tests deal with memory and CPU usage metrics. Unfortunately, I did not find any information about those metrics in the repository. The fact that Unishox2 reaches better compression rates when compared to other software is essential, indeed. However, that fact is tied to the software's core and novelty, not its performance.

Considering the tests, although someone else did the tests, the results are not published on Unishox2 repository. Instead, it points to another repository.

I do not see any problems checking both items just to end the discussion. But I am not entirely convinced that those criteria are met.
