### 2021-11-23 #3919 @gkthiruvathukal
@gradvohl Thanks for your clarification. It seems like there is nothing preventing acceptance here, but if any of these issues are blockers, please feel free to speak up. The JOSS approach is to make sure a submission results in as few open issues as possible, with the hope that the software delights users when they try to use it.

Regarding *unit tests*, my view is that if someone else did the tests--and they pass--mission accomplished!  Having said this, it would be nice if this repository shows that the tests are passing on each commit, but we don't require continuous integration (yet).

Please make sure you only check the items if you are comfortable. I can still accept, even if not all checkboxes are checked. However, I do need to give our other reviewer a chance to contribute his input before I can accept.

This constructive feedback and follow-up by the author is what JOSS is all about!!
