### 2021-11-23 #3919 @siara-cc
Hi @gradvohl , Thank you for the clarification.  So far Unishox outperforms all other technologies on compression ratio and memory usage but not speed due to complexity of logic.  I have mentioned this in the documentation and have not been able to find a better way to present it so far.

I do have Unit tests and CI pipeline in the repo that does rigorous testing with API calls and files for several languages.

I respect your assessment about these and if these are not checked based on present status I am ok with it or if you require a specific metric to pass, I will include it.  Thank you for the comprehensive review.

@gkthiruvathukal
