### 2021-11-29 #3919 @siara-cc
Hi @gradvohl  I apologise I probably was too eager in pointing out what was available and documented for unit tests and performance measurements.  Surely I was not pushing for undue credit.  I request you to kindly reassess and update the checklist reflecting the shortcomings or if you can let me know what I can do to improve it, I can at least try to earn the valuation given.
