### 2021-11-30 #3919 @linuxscout
Some review features need explanaition:
Documentation

- [ ] Installation instructions: need to be explained, but in this case, the author provided the C code, and many binding libraries for other languages. I thinks it's sufficient

- [x] Example usage:  the C code is sufficient.

- [] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified? Not given
