### 2021-11-30 #3919 @siara-cc
Hi @linuxscout Thank you for pointing out this.  It was probably not clearly mentioned in README.md, but I have made a separate section now:

[Automated tests](https://github.com/siara-cc/Unishox2#unit-tests-automated)

Unit tests are implemented in `run_unit_tests()` function of `test_unishox2.c`, which tests all the features of Unishox2, including edge cases, using 159 strings covering several languages, emojis and binary data.

Further, the CI pipeline at `.github/workflows/c-cpp.yml` automates these tests for all presets and also tests file compression for the different types of files in `sample_texts` folder. This happens whenever a commit is made to the repository.

I apologise for the delay in response.
