### 2021-12-23 #3919 @siara-cc
@gkthiruvathukal Thanks and I have done all of the above.  The tagged release version is `1.0.2` and Zenodo DOI is `10.5281/zenodo.5800408`.
Wish you merry Christmas and Happy New Year 2022.
