### 2022-01-16 #3919 @gkthiruvathukal
Thank you, @siara-cc, especially for waiting for me. I became department chair at my university at the new year, so I am just getting back to *normal*.
